pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "corteva-code-challenge"
version = "1.0.0"
requires-python = ">=3.10"
authors = [
  {name = "Rishabh Malviya", email = "rishabh.malviya@gmail.com"}
]
maintainers = [
  {name = "Rishabh Malviya", email = "rishabh.malviya@gmail.com"}
]
description = "Corteva coding challenge solution by Rishabh Malviya"
readme = "README.md"
dependencies = [
    "SQLAlchemy==2.0.35",
    "pandas==2.2.3",
    "typer==0.12.5",
    "fastapi==0.115.0",
    "fastapi-pagination==0.12.31",
    "uvicorn==0.31.0"
]

[project.optional-dependencies]
dev = [
    "jupyter==1.1.1",
    "pytest==8.3.3",
    "python-dotenv==1.0.1"
]

[project.urls]
Repository = "https://github.com/RishabhMalviya/corteva-code-challenge.git"
